+++ docs-website/src/api/plugins.ts
@@ -17,6 +17,7 @@ interface PackageJson {
 export interface Plugin {
   name: string;
   npmName: string;
+  version: string;
   slug: string;
   description: string;
   icon: string;
@@ -30,23 +31,9 @@ const pluginDirName = '../packages/';
  * Get all plugin directories starting with `vendure-plugin`
  */
 export async function getPluginDirectories(): Promise<Dirent[]> {
-  return (
-    (await readdir(pluginDirName, { withFileTypes: true }))
-      .filter((dir) => dir.isDirectory())
-      .filter((dir) => dir.name.startsWith('vendure-'))
-      // Sort alphabetically
-      .sort((a, b) => {
-        const nameA = a.name.toLowerCase();
-        const nameB = b.name.toLowerCase();
-        if (nameA < nameB) {
-          return -1;
-        }
-        if (nameA > nameB) {
-          return 1;
-        }
-        return 0;
-      })
-  );
+  return (await readdir(pluginDirName, { withFileTypes: true }))
+    .filter((dir) => dir.isDirectory())
+    .filter((dir) => dir.name.startsWith('vendure-'));
 }
 
 export async function getPlugins(): Promise<Plugin[]> {
@@ -66,15 +53,18 @@ export async function getPlugins(): Promise<Plugin[]> {
         const readmeFilePath = path.join(pluginDirName, r.name, 'README.md');
         let readme: string = await readFile(readmeFilePath, 'utf8');
         // Remove official docs link from readme
-        readme = readme.replace(/^### \[Official.*$/gm, '');
+        readme = readme.replace(/^.*Official documentation.*$/gm, '');
         // Get title from first line
         const name = readme.split('\n')[0].replace('#', '').trim();
         const readmeHtml = await parseReadme(readme);
         const nrOfDownloads = await getNrOfDownloads(packageJson.name);
-        const slug = packageJson.name.replace('@pinelab/', '');
+        const slug = packageJson.name
+          .replace('@pinelab/', '')
+          .replace('@vendure-hub/', '');
         plugins.push({
           name,
           npmName: packageJson.name,
+          version: packageJson.version,
           slug,
           description: packageJson.description,
           icon: getIcon(slug),
@@ -87,9 +77,13 @@ export async function getPlugins(): Promise<Plugin[]> {
       }
     })
   );
-  const pluginsSortedByDownloads = plugins.sort(
-    (a, b) => b.nrOfDownloads - a.nrOfDownloads
-  );
+  const pluginsSortedByDownloads = plugins.sort((a, b) => {
+    // Move vendure-hub packages to the top
+    if (a.npmName.indexOf('@vendure-hub') > -1) {
+      return -1;
+    }
+    return b.nrOfDownloads - a.nrOfDownloads;
+  });
   return pluginsSortedByDownloads;
 }