+++ packages/vendure-plugin-accept-blue/src/api/accept-blue-handler.ts
@@ -52,6 +52,13 @@ export const acceptBluePaymentHandler = new PaymentMethodHandler({
       label: [{ languageCode: LanguageCode.en, value: 'PIN' }],
       ui: { component: 'password-form-input' },
     },
+    testMode: {
+      type: 'boolean',
+      required: false,
+      defaultValue: false,
+      label: [{ languageCode: LanguageCode.en, value: 'Use test mode' }],
+      ui: { component: 'boolean-form-input' },
+    },
   },
 
   init(injector: Injector) {
@@ -83,7 +90,7 @@ export const acceptBluePaymentHandler = new PaymentMethodHandler({
       | CheckPaymentMethodInput
       | NoncePaymentMethodInput
       | SavedPaymentMethodInput;
-    const client = new AcceptBlueClient(args.apiKey, args.pin);
+    const client = new AcceptBlueClient(args.apiKey, args.pin, args.testMode);
     // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-explicit-any
     const result = await service.handlePaymentForOrder(
       ctx,