+++ packages/vendure-plugin-webhook/src/index.ts
@@ -1,4 +1,2 @@
-import { PermissionDefinition } from '@vendure/core';
-
 export * from './api/request-transformer';
 export * from './webhook.plugin';