+++ packages/vendure-plugin-accept-blue/CHANGELOG.md
@@ -1,4 +1,9 @@
-# 1.4.0 (2024-07-21)
+# 1.5.0 (2024-07-08)
+
+- Allow test mode to be set at payment method level.
+- Removed the usage of env var ACCEPT_BLUE_TEST_MODE for test mode. Use the setting on payment method.
+
+# 1.4.0 (2024-06-21)
 
 - Update Version to 2.2.6