+++ packages/vendure-plugin-google-cloud-tasks/CHANGELOG.md
@@ -1,4 +1,9 @@
-# 1.6.0 (2024-07-21)
+# 1.7.0 (2024-09-06)
+
+- Updated `@google-cloud/tasks` to `"^5.4.0"`
+- Generates a unique task name; this reduces the latency during creation of the task
+
+# 1.6.0 (2024-06-21)
 
 - Updated Vendure to 2.2.6